chore: Move Instructions M3 composable to suitable file structure
