+++ pages/src/androidTest/java/uk/gov/android/ui/pages/m3/instructions/InstructionsTest.kt
@@ -13,8 +13,6 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.Parameterized
 import uk.gov.android.ui.pages.brp.BrpInstructionsContentSection
-import uk.gov.android.ui.pages.instructions.m3.Instructions
-import uk.gov.android.ui.pages.instructions.m3.InstructionsProvider
 
 @RunWith(Parameterized::class)
 class InstructionsTest(
